test(core): enable signal queries tests with authoring
Use signal queries tests with the authoring functions and without
decorators.
